+++ server/db/disk/diskdb.go
@@ -5,6 +5,6 @@ import (
 	"fmt"
 )
 
-func InsertDiskDB(js base.SysDiskInfo,addr string,serverid int32){
+func InsertDiskDB(js base.SysDiskInfo,addr string,serverid int64){
 	fmt.Println(js)
 }
