+++ server/db/host/hostdb.go
@@ -5,6 +5,6 @@ import (
 	"fmt"
 )
 
-func InsertHostDB(js base.SysHostInfo,addr string,serverid int32){
+func InsertHostDB(js base.SysHostInfo,addr string,serverid int64){
 	fmt.Println(js)
 }
