+++ server/db/mem/memdb.go
@@ -5,6 +5,6 @@ import (
 	"fmt"
 )
 
-func InsertMemDB(js base.SysMemInfo,addr string,serverid int32){
+func InsertMemDB(js base.SysMemInfo,addr string,serverid int64){
 	fmt.Println(js)
 }
