+++ server/db/net/netdb.go
@@ -5,6 +5,6 @@ import (
 	"fmt"
 )
 
-func InsertNetDB(js base.SysNetInfo,addr string,serverid int32){
+func InsertNetDB(js base.SysNetInfo,addr string,serverid int64){
 	fmt.Println(js)
 }
